Split `AnimationFile` loading code to separate file
