+++ NAS2D/NAS2D.vcxproj
@@ -177,6 +177,7 @@
     <ClCompile Include="Renderer\RendererNull.cpp" />
     <ClCompile Include="Renderer\RendererOpenGL.cpp" />
     <ClCompile Include="Renderer\Window.cpp" />
+    <ClCompile Include="Resource\AnimationFile.cpp" />
     <ClCompile Include="Resource\AnimationFrame.cpp" />
     <ClCompile Include="Resource\AnimationSequence.cpp" />
     <ClCompile Include="Resource\AnimationSet.cpp" />
